Add submodules to trajectories
