+++ src/mrinufft/trajectories/__init__.py
@@ -18,6 +18,31 @@ from .inits import (
     initialize_3D_eccentric,
     initialize_3D_random_walk,
     initialize_3D_travelling_salesman,
+    initialize_2D_radial,
+    initialize_2D_spiral,
+    initialize_2D_vds_spiral,
+    initialize_2D_fibonacci_spiral,
+    initialize_2D_cones,
+    initialize_2D_sinusoide,
+    initialize_2D_propeller,
+    initialize_2D_rosette,
+    initialize_2D_rings,
+    initialize_2D_polar_lissajous,
+    initialize_2D_lissajous,
+    initialize_2D_waves,
+    initialize_3D_phyllotaxis_radial,
+    initialize_3D_golden_means_radial,
+    initialize_3D_wong_radial,
+    initialize_3D_park_radial,
+    initialize_3D_cones,
+    initialize_3D_floret,
+    initialize_3D_wave_caipi,
+    initialize_3D_seiffert_spiral,
+    initialize_3D_helical_shells,
+    initialize_3D_annular_shells,
+    initialize_3D_seiffert_shells,
+    initialize_3D_turbine,
+    initialize_3D_repi,
 )
 from .projection import parameterize_by_arc_length
 from .sampling import (
@@ -40,35 +65,6 @@ from .tools import (
     stack,
     stack_random,
     stack_spherically,
-)
-from .trajectory2D import (
-    initialize_2D_cones,
-    initialize_2D_vds_spiral,
-    initialize_2D_fibonacci_spiral,
-    initialize_2D_lissajous,
-    initialize_2D_polar_lissajous,
-    initialize_2D_propeller,
-    initialize_2D_radial,
-    initialize_2D_rings,
-    initialize_2D_rosette,
-    initialize_2D_sinusoide,
-    initialize_2D_spiral,
-    initialize_2D_waves,
-)
-from .trajectory3D import (
-    initialize_3D_annular_shells,
-    initialize_3D_cones,
-    initialize_3D_floret,
-    initialize_3D_golden_means_radial,
-    initialize_3D_helical_shells,
-    initialize_3D_park_radial,
-    initialize_3D_phyllotaxis_radial,
-    initialize_3D_repi,
-    initialize_3D_seiffert_shells,
-    initialize_3D_seiffert_spiral,
-    initialize_3D_turbine,
-    initialize_3D_wave_caipi,
-    initialize_3D_wong_radial,
 )
 
 __all__ = [
